Refactor code view
